Refactor topology assets into local module files
