+++ custom_components/openwrt_ubus/__init__.py
@@ -3,20 +3,16 @@
 from __future__ import annotations
 
 import asyncio
-from datetime import datetime
 import logging
 
-from aiohttp import web
 import voluptuous as vol
 
-from homeassistant.components import panel_custom, websocket_api
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_HOST, CONF_PASSWORD, CONF_USERNAME
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
 from homeassistant.helpers import config_validation as cv, device_registry as dr
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.http import HomeAssistantView
 from homeassistant.helpers.typing import ConfigType
 
 from .const import (
@@ -50,620 +46,11 @@ from .const import (
     get_config_value,
 )
 from .shared_data_manager import SharedUbusDataManager
+from .topology import async_setup_topology
 from .ubus_client import create_enhanced_extended_ubus_client
 from .security_utils import CredentialManager, safe_log_data
 
 _LOGGER = logging.getLogger(__name__)
-
-TOPOLOGY_PANEL_PATH = "openwrt-ubus-topology"
-TOPOLOGY_PANEL_MODULE_URL = "/api/openwrt_ubus/topology-panel.js"
-TOPOLOGY_WEBCOMPONENT = "openwrt-ubus-topology-panel"
-TOPOLOGY_WS_TYPE = "openwrt_ubus/topology"
-TOPOLOGY_OFFLINE_GRACE_SECONDS = 90
-
-TOPOLOGY_PANEL_JS = r'''const navigateToPath = (path) => {
-  window.history.pushState(null, "", path);
-  window.dispatchEvent(new Event("location-changed", { bubbles: true, composed: true }));
-};
-
-let visScriptPromise;
-
-const ensureVis = async () => {
-  if (window.vis?.Network) {
-    return window.vis;
-  }
-  if (!visScriptPromise) {
-    visScriptPromise = new Promise((resolve, reject) => {
-      const script = document.createElement("script");
-      script.src = "https://unpkg.com/vis-network/standalone/umd/vis-network.min.js";
-      script.onload = () => resolve(window.vis);
-      script.onerror = () => reject(new Error("Unable to load vis-network"));
-      document.head.appendChild(script);
-    });
-  }
-  return visScriptPromise;
-};
-
-class OpenWrtUbusTopologyPanel extends HTMLElement {
-  constructor() {
-    super();
-    this.attachShadow({ mode: "open" });
-    this._graphs = [];
-    this._selectedEntryId = null;
-    this._loading = false;
-    this._error = null;
-    this._resizeObserver = null;
-    this._network = null;
-    this._shouldFit = true;
-  }
-
-  set hass(hass) {
-    this._hass = hass;
-    if (!this._initialized) {
-      this._initialized = true;
-      this._render();
-      this._fetchData();
-    }
-  }
-
-  connectedCallback() {}
-
-  disconnectedCallback() {
-    this._disposeNetwork();
-  }
-
-  _startRefresh() {}
-
-  async _fetchData() {
-    if (!this._hass) {
-      return;
-    }
-    this._loading = true;
-    this._error = null;
-    this._render();
-    try {
-      const result = await this._hass.callWS({ type: "openwrt_ubus/topology" });
-      this._graphs = result.graphs || [];
-      if (!this._selectedEntryId && this._graphs.length) {
-        this._selectedEntryId = this._graphs[0].entry_id;
-      }
-      this._shouldFit = true;
-    } catch (err) {
-      this._error = err?.message || String(err);
-    } finally {
-      this._loading = false;
-      this._render();
-    }
-  }
-
-  _disposeNetwork() {
-    if (this._resizeObserver) {
-      this._resizeObserver.disconnect();
-      this._resizeObserver = null;
-    }
-    if (this._network) {
-      this._network.destroy();
-      this._network = null;
-    }
-  }
-
-  _selectedGraph() {
-    return this._graphs.find((graph) => graph.entry_id === this._selectedEntryId) || this._graphs[0] || null;
-  }
-
-  _buildVisData(graph) {
-    const style = getComputedStyle(this);
-    const apColor = style.getPropertyValue("--info-color").trim() || "#03a9f4";
-    const wirelessColor = style.getPropertyValue("--success-color").trim() || "#4caf50";
-    const wiredColor = style.getPropertyValue("--warning-color").trim() || "#ff9800";
-
-    const nodes = graph.nodes.map((node) => {
-      const details = [node.secondary];
-      if (node.signal !== undefined && node.signal !== null) {
-        details.push(`Signal: ${node.signal} dBm`);
-      }
-      if (node.confidence) {
-        details.push(`Confidence: ${node.confidence}`);
-      }
-      if (node.status) {
-        details.push(`Status: ${node.status}`);
-      }
-
-      return {
-        id: node.id,
-        label: node.label,
-        title: details.filter(Boolean).join("<br>"),
-        physics: true,
-        color:
-          node.type === "ap"
-            ? apColor
-            : node.type === "wired_device"
-              ? wiredColor
-              : node.type === "wireless_device"
-                ? wirelessColor
-                : undefined,
-        rawNode: node,
-      };
-    });
-
-    const edges = graph.edges.map((edge, index) => ({
-      id: `${edge.source}-${edge.target}-${index}`,
-      from: edge.source,
-      to: edge.target,
-      label: edge.label || undefined,
-    }));
-
-    return { nodes, edges };
-  }
-
-  async _renderNetwork() {
-    const graph = this._selectedGraph();
-    const container = this.shadowRoot?.getElementById("network");
-    if (!graph || !container) {
-      return;
-    }
-
-    try {
-      const vis = await ensureVis();
-      const data = this._buildVisData(graph);
-      const options = {
-        autoResize: true,
-        physics: {
-          enabled: true,
-          stabilization: {
-            enabled: true,
-            iterations: 150,
-            updateInterval: 25,
-            fit: true,
-          },
-          barnesHut: {
-            gravitationalConstant: -5000,
-            centralGravity: 0.18,
-            springLength: 220,
-            springConstant: 0.05,
-            damping: 0.12,
-          },
-        },
-        interaction: {
-          dragNodes: true,
-          dragView: true,
-          zoomView: true,
-          hover: true,
-          tooltipDelay: 80,
-        },
-        nodes: {
-          shadow: true,
-        },
-        edges: {
-          shadow: true,
-        },
-      };
-
-      if (!this._network) {
-        this._network = new vis.Network(container, data, options);
-        this._network.on("click", (params) => {
-          if (!params.nodes?.length) {
-            return;
-          }
-          const nodeId = params.nodes[0];
-          const clickedNode = graph.nodes.find((node) => node.id === nodeId);
-          const deviceId = clickedNode?.device_id;
-          if (deviceId) {
-            navigateToPath(`/config/devices/device/${deviceId}`);
-          }
-        });
-      } else {
-        this._network.setData(data);
-        this._network.setOptions(options);
-      }
-
-      if (this._shouldFit) {
-        this._network.fit({ animation: { duration: 250, easingFunction: "easeInOutQuad" } });
-        this._shouldFit = false;
-      }
-
-      if (!this._resizeObserver) {
-        this._resizeObserver = new ResizeObserver(() => {
-          if (this._network) {
-            this._network.redraw();
-          }
-        });
-      }
-      this._resizeObserver.disconnect();
-      this._resizeObserver.observe(container);
-    } catch (err) {
-      this._error = `Failed to load topology renderer: ${err?.message || err}`;
-      this._render();
-    }
-  }
-
-  _render() {
-    if (!this.shadowRoot) {
-      return;
-    }
-
-    this._disposeNetwork();
-
-    const graph = this._selectedGraph();
-    this.shadowRoot.innerHTML = `
-      <style>
-        :host { display:block; height:100%; }
-        .wrap { display:flex; flex-direction:column; gap:16px; padding:12px 16px 16px; height:100%; box-sizing:border-box; color:var(--primary-text-color); }
-        .header { display:flex; justify-content:space-between; align-items:center; gap:10px; }
-        .title { font-size: 22px; font-weight: 600; }
-        .toolbar { display:flex; gap:10px; align-items:center; }
-        select, .refresh { border:1px solid var(--divider-color); background:var(--card-background-color); color:var(--primary-text-color); border-radius:10px; padding:10px 12px; }
-        .refresh { cursor:pointer; }
-        .panel { flex:1; min-height:calc(100vh - 110px); display:flex; border:1px solid var(--divider-color); border-radius:18px; background:var(--card-background-color); overflow:hidden; }
-        .state { padding:32px; color:var(--secondary-text-color); }
-        #network { width:100%; height:calc(100vh - 130px); min-height:calc(100vh - 130px); }
-      </style>
-      <div class="wrap">
-        <div class="header">
-          <div class="title">OpenWrt Topology</div>
-          <div class="toolbar">
-            <select id="entrySelect">
-              ${this._graphs.map((item) => `<option value="${item.entry_id}" ${item.entry_id === this._selectedEntryId ? "selected" : ""}>${item.title}</option>`).join("")}
-            </select>
-            <button class="refresh" id="refreshBtn">Refresh</button>
-          </div>
-        </div>
-        <div class="panel">
-          ${this._loading ? `<div class="state">Loading topology...</div>` : ""}
-          ${this._error ? `<div class="state">${this._error}</div>` : ""}
-          ${!this._loading && !this._error && !graph ? `<div class="state">No OpenWrt topology data available.</div>` : ""}
-          ${!this._loading && !this._error && graph ? `<div id="network"></div>` : ""}
-        </div>
-      </div>
-    `;
-
-    const select = this.shadowRoot.getElementById("entrySelect");
-    if (select) {
-      select.addEventListener("change", (event) => {
-        this._selectedEntryId = event.target.value;
-        this._shouldFit = true;
-        this._render();
-      });
-    }
-
-    const refreshBtn = this.shadowRoot.getElementById("refreshBtn");
-    if (refreshBtn) {
-      refreshBtn.addEventListener("click", async () => this._fetchData());
-    }
-
-    if (!this._loading && !this._error && graph) {
-      this._renderNetwork();
-    }
-  }
-}
-
-customElements.define("openwrt-ubus-topology-panel", OpenWrtUbusTopologyPanel);
-'''
-
-
-class OpenWrtTopologyPanelView(HomeAssistantView):
-    """Serve the OpenWrt topology panel javascript module."""
-
-    url = TOPOLOGY_PANEL_MODULE_URL
-    name = "api:openwrt_ubus:topology_panel"
-    requires_auth = False
-
-    async def get(self, request: web.Request) -> web.Response:
-        """Return the topology panel module."""
-        return web.Response(text=TOPOLOGY_PANEL_JS, content_type="application/javascript")
-
-
-def _device_id_for_identifier(
-    device_registry: dr.DeviceRegistry,
-    identifier: tuple[str, str],
-) -> str | None:
-    """Look up a Home Assistant device id by integration identifier."""
-    device = device_registry.async_get_device(identifiers={identifier})
-    return device.id if device else None
-
-
-def _format_device_label(device_data: dict, fallback: str) -> str:
-    """Build a readable device label."""
-    hostname = device_data.get("hostname")
-    ip_address = device_data.get("ip_address")
-
-    if hostname and hostname not in ("*", fallback, fallback.upper()):
-        return hostname.split(".", 1)[0]
-    if ip_address and ip_address != "Unknown IP":
-        return ip_address
-    return fallback.replace(":", "")
-
-
-def _infer_ap_band(ap_data: dict) -> str | None:
-    """Infer Wi-Fi band from AP info."""
-    channel = ap_data.get("channel")
-    frequency = ap_data.get("frequency")
-
-    try:
-        if frequency is not None:
-            freq = int(float(frequency))
-            if 2400 <= freq < 2500:
-                return "2.4G"
-            if 4900 <= freq < 5900:
-                return "5G"
-            if 5900 <= freq < 7100:
-                return "6G"
-    except (TypeError, ValueError):
-        pass
-
-    try:
-        if channel is not None:
-            ch = int(channel)
-            if 1 <= ch <= 14:
-                return "2.4G"
-            if 36 <= ch <= 177:
-                return "5G"
-    except (TypeError, ValueError):
-        pass
-
-    return None
-
-
-def _normalize_ap_identifier(value: str | None) -> str:
-    """Normalize AP identifiers across iwinfo and hostapd naming."""
-    if not value:
-        return ""
-
-    normalized = value.strip().lower()
-    prefixes = (
-        "hostapd.",
-        "hostapd:",
-        "phy",
-    )
-
-    if normalized.startswith("hostapd.") or normalized.startswith("hostapd:"):
-        normalized = normalized.split(".", 1)[-1].split(":", 1)[-1]
-
-    if "." in normalized and normalized.startswith("radio"):
-        normalized = normalized.split(".", 1)[-1]
-
-    return normalized
-
-
-def _merge_with_previous_topology(
-    previous_graph: dict | None,
-    nodes: list[dict],
-    edges: list[dict],
-) -> tuple[list[dict], list[dict]]:
-    """Keep recently missing devices visible as offline in topology."""
-    if not previous_graph:
-        return nodes, edges
-
-    now_ts = datetime.now().timestamp()
-    existing_node_ids = {node["id"] for node in nodes}
-    previous_nodes = {node["id"]: node for node in previous_graph.get("nodes", [])}
-    previous_edges = previous_graph.get("edges", [])
-
-    merged_nodes = list(nodes)
-    merged_edges = list(edges)
-    edge_keys = {(edge["source"], edge["target"], edge.get("kind")) for edge in edges}
-
-    for node_id, old_node in previous_nodes.items():
-        if node_id in existing_node_ids:
-            continue
-        if old_node.get("type") in ("router", "ap"):
-            continue
-
-        last_seen = old_node.get("topology_last_seen")
-        if last_seen is None or now_ts - last_seen > TOPOLOGY_OFFLINE_GRACE_SECONDS:
-            continue
-
-        offline_node = dict(old_node)
-        offline_node["status"] = "offline"
-        offline_node["connected"] = False
-        merged_nodes.append(offline_node)
-
-        for edge in previous_edges:
-            if edge.get("target") != node_id:
-                continue
-            edge_key = (edge.get("source"), edge.get("target"), edge.get("kind"))
-            if edge_key in edge_keys:
-                continue
-            offline_edge = dict(edge)
-            offline_edge["active"] = False
-            merged_edges.append(offline_edge)
-            edge_keys.add(edge_key)
-
-    return merged_nodes, merged_edges
-
-
-async def _build_entry_topology(hass: HomeAssistant, entry: ConfigEntry) -> dict:
-    """Build topology data for a single config entry."""
-    data_manager = hass.data.get(DOMAIN, {}).get(f"data_manager_{entry.entry_id}")
-    previous_graph = hass.data.get(DOMAIN, {}).get("topology_cache", {}).get(entry.entry_id)
-    if data_manager is None:
-        return {
-            "entry_id": entry.entry_id,
-            "title": entry.title,
-            "host": entry.data[CONF_HOST],
-            "nodes": [],
-            "edges": [],
-            "summary": {"routers": 0, "aps": 0, "wireless": 0, "wired": 0, "online": 0},
-        }
-
-    combined = await data_manager.get_combined_data(["ap_info", "device_statistics", "wired_devices"])
-    ap_info = combined.get("ap_info", {}) or {}
-    wireless_devices = combined.get("device_statistics", {}) or {}
-    wired_devices = combined.get("wired_devices", {}) or {}
-
-    device_registry = dr.async_get(hass)
-    host = entry.data[CONF_HOST]
-    router_node_id = f"router:{entry.entry_id}"
-    now_ts = datetime.now().timestamp()
-
-    nodes: list[dict] = [
-        {
-            "id": router_node_id,
-            "type": "router",
-            "label": entry.title or host,
-            "secondary": host,
-            "status": "online",
-            "device_id": _device_id_for_identifier(device_registry, (DOMAIN, host)),
-        }
-    ]
-    edges: list[dict] = []
-
-    ap_node_ids: dict[str, str] = {}
-    ap_node_ids_normalized: dict[str, str] = {}
-    for ap_device, ap_data in sorted(ap_info.items()):
-        ap_node_id = f"ap:{entry.entry_id}:{ap_device}"
-        ap_node_ids[ap_device] = ap_node_id
-        ap_node_ids_normalized[_normalize_ap_identifier(ap_device)] = ap_node_id
-        ssid = ap_data.get("ssid") or ap_data.get("device_name") or ap_device
-        band = _infer_ap_band(ap_data)
-        mode = ap_data.get("mode") or "AP"
-        label = f"{ssid} {band}" if band else ssid
-        secondary = f"{ap_device} | {mode}"
-        if band:
-            secondary = f"{secondary} | {band}"
-        nodes.append(
-            {
-                "id": ap_node_id,
-                "type": "ap",
-                "label": label,
-                "secondary": secondary,
-                "status": "online",
-                "topology_last_seen": now_ts,
-                "device_id": _device_id_for_identifier(device_registry, (DOMAIN, f"{host}_ap_{ap_device}")),
-            }
-        )
-        edges.append(
-            {
-                "source": router_node_id,
-                "target": ap_node_id,
-                "kind": "uplink",
-                "label": ap_data.get("channel"),
-                "active": True,
-            }
-        )
-
-    wireless_count = 0
-    wired_count = 0
-    online_count = 1
-
-    for mac, device_data in sorted(wireless_devices.items()):
-        node_id = f"device:{entry.entry_id}:{mac}"
-        connected = bool(device_data.get("connected", False))
-        status = "online" if connected else "offline"
-        if connected:
-            wireless_count += 1
-            online_count += 1
-
-        ap_device = device_data.get("ap_device")
-        target_id = ap_node_ids.get(ap_device)
-        if not target_id:
-            target_id = ap_node_ids_normalized.get(_normalize_ap_identifier(ap_device))
-        if not target_id:
-            target_id = router_node_id
-        signal = device_data.get("signal")
-        nodes.append(
-            {
-                "id": node_id,
-                "type": "wireless_device",
-                "label": _format_device_label(device_data, mac),
-                "secondary": mac,
-                "status": status,
-                "connected": connected,
-                "signal": signal,
-                "topology_last_seen": now_ts,
-                "device_id": _device_id_for_identifier(device_registry, (DOMAIN, mac)),
-            }
-        )
-        edges.append(
-            {
-                "source": target_id,
-                "target": node_id,
-                "kind": "wireless",
-                "label": f"{signal} dBm" if signal is not None else None,
-                "active": connected,
-            }
-        )
-
-    for mac, device_data in sorted(wired_devices.items()):
-        node_id = f"device:{entry.entry_id}:wired:{mac}"
-        connected = bool(device_data.get("connected", False))
-        status = "online" if connected else "offline"
-        if connected:
-            wired_count += 1
-            online_count += 1
-
-        interface = device_data.get("interface") or "LAN"
-        confidence = device_data.get("confidence")
-        nodes.append(
-            {
-                "id": node_id,
-                "type": "wired_device",
-                "label": _format_device_label(device_data, mac),
-                "secondary": f"{mac} | {interface}",
-                "status": status,
-                "connected": connected,
-                "confidence": confidence,
-                "topology_last_seen": now_ts,
-                "device_id": _device_id_for_identifier(device_registry, (DOMAIN, mac)),
-            }
-        )
-        edges.append(
-            {
-                "source": router_node_id,
-                "target": node_id,
-                "kind": "wired",
-                "label": interface,
-                "active": connected,
-            }
-        )
-
-    nodes, edges = _merge_with_previous_topology(previous_graph, nodes, edges)
-
-    graph = {
-        "entry_id": entry.entry_id,
-        "title": entry.title or host,
-        "host": host,
-        "nodes": nodes,
-        "edges": edges,
-        "summary": {
-            "routers": 1,
-            "aps": len(ap_node_ids),
-            "wireless": wireless_count,
-            "wired": wired_count,
-            "online": online_count,
-        },
-    }
-
-    hass.data.setdefault(DOMAIN, {}).setdefault("topology_cache", {})[entry.entry_id] = graph
-    return graph
-
-
-@websocket_api.websocket_command(
-    {
-        vol.Required("type"): TOPOLOGY_WS_TYPE,
-        vol.Optional("entry_id"): str,
-    }
-)
-@websocket_api.async_response
-async def websocket_get_topology(
-    hass: HomeAssistant,
-    connection: websocket_api.ActiveConnection,
-    msg: dict,
-) -> None:
-    """Return topology graph data."""
-    entry_id = msg.get("entry_id")
-    entries = hass.config_entries.async_entries(DOMAIN)
-    if entry_id:
-        entries = [entry for entry in entries if entry.entry_id == entry_id]
-
-    graphs = []
-    for entry in entries:
-        try:
-            graphs.append(await _build_entry_topology(hass, entry))
-        except Exception as exc:
-            _LOGGER.error("Failed building topology for %s: %s", entry.entry_id, exc)
-
-    connection.send_result(msg["id"], {"graphs": graphs})
 
 CONFIG_SCHEMA = vol.Schema(
     {
@@ -688,19 +75,7 @@ CONFIG_SCHEMA = vol.Schema(
 async def async_setup(hass: HomeAssistant, config: ConfigType) -> bool:
     """Set up the openwrt ubus component."""
     hass.data.setdefault(DOMAIN, {})
-
-    if not hass.data[DOMAIN].get("topology_registered"):
-        hass.http.register_view(OpenWrtTopologyPanelView)
-        websocket_api.async_register_command(hass, websocket_get_topology)
-        await panel_custom.async_register_panel(
-            hass,
-            frontend_url_path=TOPOLOGY_PANEL_PATH,
-            webcomponent_name=TOPOLOGY_WEBCOMPONENT,
-            module_url=TOPOLOGY_PANEL_MODULE_URL,
-            require_admin=True,
-            config={"domain": DOMAIN, "title": "OpenWrt Topology"},
-        )
-        hass.data[DOMAIN]["topology_registered"] = True
+    await async_setup_topology(hass)
 
     if DOMAIN not in config:
         return True
